Road Lane-RF DETR/requirements.txt
ultralytics
opencv-python
numpy
Pillow
matplotlib
sahi
inference_sdk